# doc/api-docs-smart/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PinShop Server",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
...
  "paths": {
...
    "/api/v1/pay/create": {
      "post": {
        "summary": "Create payment",
        "deprecated": false,
        "tags": [
          "PaymentController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/45be7e518bfcbb25ac78da375d6e2249"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"userId\": 0,\n  \"orderId\": 0,\n  \"payType\": \"\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/e67c272091d2e4d4ecedc647dbe52359"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {\n    \"paymentId\": \"\",\n    \"orderId\": 0,\n    \"amount\": 0,\n    \"payType\": \"\",\n    \"payUrl\": \"\"\n  },\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-pay-create-POST",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/pay/callback/wechat": {
      "post": {
        "summary": "WeChat payment callback",
        "deprecated": false,
        "tags": [
          "PaymentController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/36d31e077ebfca7ac879057e5683a2a0"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"paymentId\": \"\",\n  \"transactionId\": \"\",\n  \"amount\": 0,\n  \"status\": \"\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/94e3d32fa43fbe1c67c36d27ee7aaa69"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"returnCode\": \"\",\n  \"returnMsg\": \"\"\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-pay-callback-wechat-POST",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/pay/callback/alipay": {
      "post": {
        "summary": "Alipay payment callback",
        "deprecated": false,
        "tags": [
          "PaymentController"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/36d31e077ebfca7ac879057e5683a2a0"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"paymentId\": \"\",\n  \"transactionId\": \"\",\n  \"amount\": 0,\n  \"status\": \"\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/94e3d32fa43fbe1c67c36d27ee7aaa69"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"returnCode\": \"\",\n  \"returnMsg\": \"\"\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-pay-callback-alipay-POST",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
    "/api/v1/pay/status": {
      "get": {
        "summary": "Get payment status",
        "deprecated": false,
        "tags": [
          "PaymentController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/71eaafec35e657932b499a3d718d7ddb"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {\n    \"orderId\": 0,\n    \"status\": 0,\n    \"statusText\": \"\",\n    \"payAmount\": 0,\n    \"payTime\": \"yyyy-MM-dd HH:mm:ss\"\n  },\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/da0c932b0e6d3b90859201037a7929ee"
                },
                "examples": {
                  "json": {
                    "summary": "response example",
                    "value": "{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1709823456789\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-pay-status-GET",
        "parameters": [
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "user ID",
            "required": true,
            "example": "0",
            "name": "userId"
          },
          {
            "schema": {
              "format": "int64",
              "type": "number"
            },
            "in": "query",
            "description": "order ID",
            "required": true,
            "example": "0",
            "name": "orderId"
          },
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "header",
            "description": "认证Token (Bearer {jwt_token})",
            "required": false,
            "name": "Authorization"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "71eaafec35e657932b499a3d718d7ddb": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "$ref": "#/components/schemas/2d2880930c48e089523a4522ce056faa"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "94e3d32fa43fbe1c67c36d27ee7aaa69": {
        "properties": {
          "returnCode": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "returnMsg": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "2d2880930c48e089523a4522ce056faa": {
        "properties": {
          "orderId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "No comments found.",
            "type": "integer",
            "example": "0"
          },
          "statusText": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "payAmount": {
            "description": "No comments found.",
            "example": "0"
          },
          "payTime": {
            "format": "string",
            "description": "No comments found.",
            "type": "string",
            "example": "yyyy-MM-dd HH:mm:ss"
          }
        }
      },
...
      "e67c272091d2e4d4ecedc647dbe52359": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "$ref": "#/components/schemas/a1bf2f9a2cedc5f8be18a991b7d87c63"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
      "36d31e077ebfca7ac879057e5683a2a0": {
        "properties": {
          "paymentId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "transactionId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "amount": {
            "description": "No comments found.",
            "example": "0"
          },
          "status": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "a1bf2f9a2cedc5f8be18a991b7d87c63": {
        "properties": {
          "paymentId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "orderId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "amount": {
            "description": "No comments found.",
            "example": "0"
          },
          "payType": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "payUrl": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "45be7e518bfcbb25ac78da375d6e2249": {
        "properties": {
          "userId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "orderId": {
            "format": "int64",
            "description": "No comments found.",
            "type": "number",
            "example": "0"
          },
          "payType": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "da0c932b0e6d3b90859201037a7929ee": {
        "properties": {
          "code": {
            "format": "int32",
            "description": "业务状态码",
            "type": "integer",
            "example": "200"
          },
          "message": {
            "format": "string",
            "description": "响应消息",
            "type": "string",
            "example": "success"
          },
          "data": {
            "description": "业务数据(object)",
            "type": "object"
          },
          "timestamp": {
            "format": "int64",
            "description": "响应时间戳(毫秒)",
            "type": "number",
            "example": "1709823456789"
          }
        }
      },
...
    }
  }
}